refactor: split audio_engine.py (919 LOC) into subpackage per AGENTS.md Rule 14
- Create mcp_video/audio_engine/ subpackage with 4 modules:
  core.py (323) - waveform generators, filters, reverb, wav I/O
  synthesis.py (327) - audio_synthesize, audio_preset
  sequencing.py (252) - audio_sequence, audio_compose, audio_effects
  __init__.py (128) - add_generated_audio + re-exports
- Preserves backward-compatible import path
  (from mcp_video.audio_engine import audio_synthesize, ...)

Tests: 773 passed, 9 skipped, 2 xpassed
